requirements.txt
certifi==2018.10.15
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.2
django-cors-headers==2.4.0
django-filter==2.0.0
django-filters==0.2.1
django-heroku==0.3.1
djangorestframework==3.9.0
gunicorn==19.9.0
idna==2.7
isodate==0.6.0
numpy==1.15.3
psycopg2==2.7.6.1
python-dateutil==2.7.5
python-dotenv==0.10.1
pytz==2018.7
requests==2.20.1
semver==2.8.1
six==1.11.0
urllib3==1.24.1
whitenoise==4.1.2
